+++ CHANGELOG.md
@@ -7,8 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## latest
 * Add studioLink and studioLinkEnabled
-* Update sass-loader to get rid of warnings
+* Add Template Element and Template Parts
 * Make all properties available as scss variables during the build
+* Update sass-loader to get rid of warnings
 
 
 ## [2.0.0] - 09.10.2025